+++ R/core_functions.R
@@ -1,0 +1,45 @@
+#' @export
+#' @import twc
+twc::crop_data
+
+#' @export
+twc::csi
+
+#' @export
+twc::far
+
+#' @export
+twc::fldmean
+
+#' @export
+twc::infoNC
+
+#' @export
+twc::muldpm
+
+#' @export
+twc::nse
+
+#' @export
+twc::pRecipe_masks
+
+#' @export
+twc::pod
+
+#' @export
+twc::remap
+
+#' @export
+twc::saveNC
+
+#' @export
+twc::subset_data
+
+#' @export
+twc::tabular
+
+#' @export
+twc::trend
+
+#' @export
+twc::yearstat