+++ R/download_cpc.R
@@ -13,6 +13,7 @@
 #' }
 #' @param time_res a character string with the desired time resolution. Suitable options are:
 #' \itemize{
+#' \item{"daily",}
 #' \item{"monthly",}
 #' \item{"yearly".}
 #' }
@@ -28,9 +29,10 @@ download_cpc <- function(folder_path = ".", domain = "raw", time_res = "monthly"
   } else {
     warning(paste0('The ', domain, ' domain is not available'))
   }
-  zenodo_base <- "https://zenodo.org/record/7808922/files/"
+  zenodo_base <- "https://zenodo.org/record/14290970/files/"
+  if (time_res == "daily") {zenodo_base <- "https://zenodo.org/record/14290969/files/"}
   zenodo_end <- "?download=1"
-  file_name <- paste0("cpc_tp_mm_", domain, "_197901_202208_025_", time_res, ".nc")
+  file_name <- paste0("cpc-global_tp_mm_", domain, "_197901_202309_025_", time_res, ".nc")
   file_url <- paste0(zenodo_base, file_name, zenodo_end)
   file_destination <- paste(folder_path, file_name, sep = "/")
   try(download.file(file_url, file_destination, mode = "wb"), silent = TRUE)