+++ R/download_cru_ts.R
@@ -28,9 +28,9 @@ download_cru_ts <- function(folder_path = ".", domain = "raw", time_res = "month
   } else {
     warning(paste0('The ', domain, ' domain is not available'))
   }
-  zenodo_base <- "https://zenodo.org/record/7808922/files/"
+  zenodo_base <- "https://zenodo.org/record/14290970/files/"
   zenodo_end <- "?download=1"
-  file_name <- paste0("cru-ts_tp_mm_", domain, "_190101_202112_025_", time_res, ".nc")
+  file_name <- paste0("cru-ts-v4-08_tp_mm_", domain, "_190101_202312_025_", time_res, ".nc")
   file_url <- paste0(zenodo_base, file_name, zenodo_end)
   file_destination <- paste(folder_path, file_name, sep = "/")
   try(download.file(file_url, file_destination, mode = "wb"), silent = TRUE)