+++ R/download_era5.R
@@ -24,7 +24,8 @@ download_era5 <- function(folder_path = ".", domain = "raw", time_res = "monthly
   options(timeout = 6000)
   on.exit(options(old_options))
   if (domain == "raw"){domain <- "global"}
-  zenodo_base <- "https://zenodo.org/record/7808922/files/"
+  zenodo_base <- "https://zenodo.org/record/14290970/files/"
+  if (domain == "ocean") {zenodo_base <- "https://zenodo.org/record/14290971/files/"}
   zenodo_end <- "?download=1"
   file_name <- paste0("era5_tp_mm_", domain, "_195901_202112_025_", time_res, ".nc")
   file_url <- paste0(zenodo_base, file_name, zenodo_end)