+++ R/download_gsmap.R
@@ -1,6 +1,6 @@
-#' UDEL data downloader
+#' GPM_IMERG data downloader
 #'
-#' Function for downloading UDEL v5.01.
+#' Function for downloading GPM IMERGM v06.
 #'
 #' @importFrom utils download.file
 #' @param folder_path a character string with the path where the data will be downloaded.
@@ -13,24 +13,23 @@
 #' }
 #' @param time_res a character string with the desired time resolution. Suitable options are:
 #' \itemize{
+#' \item{"daily",}
 #' \item{"monthly",}
 #' \item{"yearly".}
 #' }
 #' @return No return value, called to download the data set.
 #' @keywords internal
 
-download_udel <- function(folder_path = ".", domain = "raw", time_res = "monthly"){
+download_gsmap <- function(folder_path = ".", domain = "raw", time_res = "monthly"){
   old_options <- options()
   options(timeout = 6000)
   on.exit(options(old_options))
-  if (domain == "raw" | domain == "land"){
-    domain <- "land"
-  } else {
-    warning(paste0('The ', domain, ' domain is not available'))
-  }
-  zenodo_base <- "https://zenodo.org/record/7808922/files/"
+  if (domain == "raw"){domain <- "global"}
+  zenodo_base <- "https://zenodo.org/record/14290970/files/"
+  if (domain == "ocean") {zenodo_base <- "https://zenodo.org/record/14290971/files/"}
+  if (time_res == "daily") {zenodo_base <- "https://zenodo.org/record/14290969/files/"}
   zenodo_end <- "?download=1"
-  file_name <- paste0("udel_tp_mm_", domain, "_190001_201712_025_", time_res, ".nc")
+  file_name <- paste0("gsmap-v8_tp_mm_", domain, "_199801_202306_025_", time_res, ".nc")
   file_url <- paste0(zenodo_base, file_name, zenodo_end)
   file_destination <- paste(folder_path, file_name, sep = "/")
   try(download.file(file_url, file_destination, mode = "wb"), silent = TRUE)