+++ R/globals.R
@@ -1,3 +1,3 @@
-utils::globalVariables(c(".", "A", "B", "C", "D", "dataset", "forecast", "idx",
-                         "lat", "lon", "observed", "seasons", "t1", "t2",
-                         "value"))
+utils::globalVariables(c(".", "A", "B", "C", "coord_id", "D", "dataset",
+                         "forecast", "idx", "lat", "lon", "observed", "seasons",
+                         "t1", "t2", "value"))