+++ R/plot_taylor.R
@@ -3,13 +3,14 @@
 #' Convenient and aesthetic visualization of data in a Taylor diagram.
 #' 
 #' @details
-#' `x` columns should be named: "lon", "lat", "date", "value", "dataset", and "source". The last two columns are added using the \code{\link{label}}.
+#' `x` columns should be named: "lon", "lat", "date", "value", "dataset", and "source".
 #' 
-#' `y` columns should be named: "lon", "lat", "date", "value", "dataset", and "source". The last two columns are added using the \code{\link{label}}.
+#' `y` columns should be named: "lon", "lat", "date", "value", "dataset", and "source".
 #' 
 #' `groups` character to define panels. Suitable options are:
 #' \itemize{
-#' \item "source" (default)
+#' \item "default" (only one panel)
+#' \item "source"
 #' \item "seasons" (only works properly with monthly data)
 #' }
 #' 
@@ -25,7 +26,7 @@
 #' @return plot object
 #' @export
 
-plot_taylor <- function(x, y, groups = "source", ...){
+plot_taylor <- function(x, y, groups = "default", ...){
   text_obs <- y$dataset[1]
   y <- y[, .(obs = value), .(date)]
   precip <- merge(y, x, by = "date")